refactor: map store folder structure
