+++ packages/editor/hooks/useScale.ts
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { useSelector } from 'react-redux';
-import { selectTileSize } from '../stores/map/selectors';
+import { selectTileSize } from '../store/map/selectors';
 
 export const useScale = () => {
     const tileSize = useSelector(selectTileSize);
