+++ packages/editor/pages/_app.tsx
@@ -1,13 +1,13 @@
-import React from "react";
+import React from 'react';
 import { Provider } from 'react-redux';
-import { store } from '../stores/map';
+import store from '../store';
 
 function MyApp({ Component, pageProps }) {
-  return (
-    <Provider store={store}>
-      <Component {...pageProps} />
-    </Provider>
-  );
+    return (
+        <Provider store={store}>
+            <Component {...pageProps} />
+        </Provider>
+    );
 }
 
 export default MyApp;
