+++ packages/editor/store/helpers.ts
@@ -1,0 +1,4 @@
+export const createAction = <Payload extends object, T extends string>(type: T) => (payload: Payload) => ({
+    type,
+    payload,
+});
