+++ packages/editor/store/index.ts
@@ -1,0 +1,6 @@
+import { createStore } from 'redux';
+import reducer from './map/reducer';
+
+const store = createStore(reducer);
+
+export default store;
