--- packages/editor/stores/map/types.ts
@@ -1,3 +1,0 @@
-import { config } from ".";
-
-export type State = typeof config;
